Learning some ICED.

+++ src/planet.rs
@@ -1,6 +1,6 @@
 #[derive(Debug)]
 pub struct Planet {
-    pub pos_x: f64,
-    pub pos_y: f64,
+    pub pos_x: f32,
+    pub pos_y: f32,
     pub name: String,
 }
